Move routers from main.py to router.py in 'food_items' directory
